Add files via upload
v1.0.0 - Basic Tracking
Info on the mod's features

/bn toggle: Starts and stops the bazaar helper. Do this before starting your flipping
/bn dump: Dumps the buy orders that are stored in .minecraft/logs/latest.log
/bn api (key): Sets your api key. This is necessary for doing /bn toggle.
- Example: /bn api xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
/bn empty: Empties the list of orders that are being stored. This might be necessary in case of some bugs.
What this version introduced

If your buy order isn't on the top of the list, you will be notified in chat.
If your sell offer isn't on the top of the list, you will be notified in chat.
Upcoming Features

Making a version of the bot (list top items) inside the game so you don't need to switch to discord
More Info

To install, make sure you have Forge for Minecraft 1.8.9. Just put the mod in the mods folder.